+++ R/zzz.R
@@ -2,5 +2,6 @@ utils::globalVariables(c(
   "membership", "registered", "casual", "latitude", "longitude",
   "Weekday", "holiday", "valid", "wxcodes", "skyc1", "tmpf", "sknt",
   "relh", "alti", "vsby", "weather_text", "Temp", "Wind", "Humidity",
-  "Barometer", "Visibility", "archive_id", "archive_name", "total_rentals"
+  "Barometer", "Visibility", "archive_id", "archive_name", "total_rentals",
+  "archive", "name", "city", "system_data_url", "started_at"
 ))